% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{discretizaComCutPoints}
\alias{discretizaComCutPoints}
\title{Discretiza coluna usando valores de cut points}
\usage{
discretizaComCutPoints(coluna, cut.points)
}
\arguments{
\item{coluna}{coluna a ser discretizada.}

\item{cut.points}{pontos de corte para separar valores da coluna.}
}
\value{
factor com coluna discretizada por cut.points.
}
\description{
Esta função transforma uma coluna numérica em factor, separando os valores da
coluna a partir dos cut points definidos.
}
\details{
Esta função transforma uma coluna numérica em factor, separando os valores da
coluna a partir dos cut points definidos.
Portanto, esta função não define os cut.points, ela só os aplica em uma coluna
para obter o resultado discretizado.
Os limites inferior e superio usados são -Inf e Inf.
Os intervalos são sempre definidos excluindo o menor valor e incluindo o maior valor.
}
\examples{
# Dados do exemplo
xcoluna <- c(rep(1,10), rep(100,10), rep(1000,10))
xclasse <- c(rep(0,10), rep(1,10), rep(0,10))
dados <- cbind(xcoluna, xclasse)
# Realizar discretização via CAIM para obter cut.points
# É método supervisionado, portanto, discretiza xcoluna "vendo" xclasse
require(discretization)
discret <- disc.Topdown(dados, method = 1) # CAIM
# Extrair cut.points
cut.points <- discret$cutp[[1]]
# Discretizar xcoluna com cut.points
xcoluna_disc <- discretizaComCutPoints(xcoluna, cut.points)
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{limpaCaracteres}, \code{getDFcomDummyCols}
}

