% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{getDFcomDummyCols}
\alias{getDFcomDummyCols}
\title{Transforma colunas character/factor em colunas dummy}
\usage{
getDFcomDummyCols(df.completo, cols.to.dummy)
}
\arguments{
\item{df.completo}{data frame completo, com todas as colunas.}

\item{cols.to.dummy}{vetor com números indicando quais colunas de df.completo 
deverão ser transformadas em dummy.}
}
\value{
df.completo combinado (cbind) com colunas transformadas em dummy.
}
\description{
Esta função transforma colunas selecionadas em dummy e retorna o data frame
original combinado (cbind) com as colunas dummy.
}
\details{
Esta função transforma em dummy as colunas do df.completo nas posições definidas em 
cols.to.dummy e retorna o data frame original df.completo combinado (cbind) com as 
colunas transformadas em dummy.
}
\examples{
# Obter DFRFB junto com colunas 3,6 e 11 trasnformadas em dummy
DFRFB_com_dummy <- getDFcomDummyCols(DFRFB, c(3,6,11))
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{limpaCaracteres}, \code{discretizaComCutPoints}
}

