% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{insertDataIntoDB}
\alias{insertDataIntoDB}
\title{Inserir dados em tabela de SGBD}
\usage{
insertDataIntoDB(nomeConexaoODBC, dadosNOVOS, nomeTabela,
  verboseInsert = FALSE, showSuccessMessage = TRUE)
}
\arguments{
\item{nomeConexaoODBC}{nome do Data Source configurado para o SGBD desejado.}

\item{dadosNOVOS}{data frame com dados a serem inseridos. Deve conter colunas com
exatamente os mesmos nomes que as colunas da tabela onde dados serão inseridos 
(case sensitive).}

\item{nomeTabela}{nome da tabela do SGBD onde dados serão inseridos.}

\item{verboseInsert}{flag que define se deve exibir mensagens de inserção do SGBD.}

\item{showSuccessMessage}{flag que define se a mensagem de sucesso deverá ser exibida.}
}
\value{
mensagem de erro ou sucesso.
}
\description{
Esta função se conecta a um SGBD definido por um Data Source configurado e 
usa usuário/senha obtidos em arquivo database_logins.txt e insere dados passados
na tabela escolhida.
}
\details{
Esta função inicialmente lê o arquivo database_logins.txt que deve estar na
mesma pasta do Working Directory (definido com setwd) e obtém usuário e senha 
para o Data Source indicado como argumento. 
Em seguida, insere data frame dadosNOVOS na tabela nomeTabela.
O objeto dadosNOVOS deve conter colunas com exatamente os mesmos nomes que as colunas 
da tabela onde dados serão inseridos (case sensitive).
Para flag verboseInsert setada, as mensagens do SGBD para o insert são exibidas.
Para flag showSuccessMessage setada, havendo sucesso na consulta, a mensagem
de sucesso padrão é exibida. Caso sejam executados vários inserts em sequência,
recomenda-se definir showSuccessMessage como FALSE, pois assim somente aparece
mensagem em caso de ERRO em algum insert.
}
\examples{
# Inserir data frama df_NatResp usando data source sed-die-bd1-c
# na tabela dbo.natresp localizada no dw_mara_stage
insertDataIntoDB('sed-die-bd1-c', df_NatResp, 'natresp')
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{runSQLonDB}, \code{instrucoes.SQLServer}, \code{instrucoes.MySQL}
}

