% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{instrucoes.MySQL}
\alias{instrucoes.MySQL}
\title{Instruções para uso de MySQL}
\usage{
instrucoes.MySQL()
}
\value{
mensagem com instruções.
}
\description{
Esta função informa passos de como ficar pronto para realizar consultas a um
SGBD MySQL (versões 3 ou 5).
}
\details{
Esta função não recebe argumentos e retorna um texto que informa os passos para
habilitar o uso uma máquina Windows para realização de consultas e inserções
de dados em um SGBD MySQL (versões 3 ou 5).
}
\examples{
# Exibir mensagem com instruções
instrucoes.MySQL()
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{instrucoes.SQLServer}, \code{insertDataIntoDB}, \code{runSQLonDB}
}

