% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{instrucoes.SQLServer}
\alias{instrucoes.SQLServer}
\title{Instruções para uso de SQL Server}
\usage{
instrucoes.SQLServer()
}
\value{
mensagem com instruções.
}
\description{
Esta função informa passos de como ficar pronto para realizar consultas a um
SGBD SQL Server.
}
\details{
Esta função não recebe argumentos e retorna um texto que informa os passos para
habilitar o uso uma máquina Windows para realização de consultas e inserções
de dados em um SGBD SQL Server.
}
\examples{
# Exibir mensagem com instruções
instrucoes.SQLServer()
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{instrucoes.MySQL}, \code{insertDataIntoDB}, \code{runSQLonDB}
}


