% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{limpaCaracteres}
\alias{limpaCaracteres}
\title{Limpar caracteres de um vetor/coluna}
\usage{
limpaCaracteres(coluna, removeEspacos = FALSE, metodoTRANSLIT = TRUE)
}
\arguments{
\item{coluna}{vetor que deve ser limpo.}

\item{removeEspacos}{flag que define se deve remover completamente os espaços em branco.}

\item{metodoTRANSLIT}{flag que define se será usado método ASCII/TRANSLIT para limpeza de acentos.
Caso seja FALSE usará conversão de latin para ASCII como método para limpeza de acentos.}
}
\value{
vetor limpo.
}
\description{
Esta função realiza diversas limpezas para um vetor/coluna do tipo character.
}
\details{
Esta função realiza as seguintes limpezas: caso flag removeEspacos seja TRUE remove
completamente todos os espaços do vetor; remove acentuação, por um de dois métodos 
(ASCII//TRANSLIT ou latin para ASCII) escolhido pelo flag metodoTRANSLIT; remove
espaços que ficam no começo ou fim; remove pontuações; e transforma textos em maiúsculo.
}
\examples{
# Limpar coluna nm_pessoa do data frama DFRFB
# Não removendo espaços em branco e usando método TRANSLIT de remoção de acentuação
DFRFB$nm_pessoa <- limpaCaracteres(DFRFB$nm_pessoa)
# Caso não funcione a limpeza com método TRANSLIT, tentar conversão latin1 para ASCII
DFRFB$nm_pessoa <- limpaCaracteres(DFRFB$nm_pessoa, metodoTRANSLIT=FALSE)
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{getDFcomDummyCols}, \code{discretizaComCutPoints}
}

