% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mara_Minigenom}
\alias{Mara_Minigenom}
\title{Minigenome Data}
\format{
A data frame with 17 variables:
\describe{
\item{\code{ID}}{Replicate of the Experiment}
\item{\code{WT}}{relative Minigenom Aktivität des Wildtyp NP}
\item{\code{V199A}}{relative Minigenom Aktivität der V199A NP Mutante}
\item{\code{F201A}}{relative Minigenom Aktivität der F201A NP Mutante}
\item{\code{V199F201AA}}{relative Minigenom Aktivität der V199A/F201A NP Mutante}
\item{\code{R174A}}{relative Minigenom Aktivität der R174A NP Mutante}
\item{\code{R174I176V178AAA}}{relative Minigenom Aktivität der R174A I176A V178A NP Mutante}
\item{\code{R174I176V178V199F201AAAAA}}{relative Minigenom Aktivität der R174A I176A V178A V199A/F201A NP Mutante}
\item{\code{S581A}}{relative Minigenom Aktivität der S581A NP Mutante}
\item{\code{S581D}}{relative Minigenom Aktivität der S581D NP Mutante}
\item{\code{S587A}}{relative Minigenom Aktivität der S587A NP Mutante}
\item{\code{S587D}}{relative Minigenom Aktivität der S587D NP Mutante}
\item{\code{S647A}}{relative Minigenom Aktivität der S647A NP Mutante}
\item{\code{S647D}}{relative Minigenom Aktivität der S647D NP Mutante}
\item{\code{ohneNP}}{relative Minigenom Aktivität der ohne NP Mutante}
}
}
\usage{
Mara_Minigenom
}
\description{
A dataset of different Ebola virus NP-Mutants in minigenome assay
}
\keyword{datasets}
