% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{runSQLonDB}
\alias{runSQLonDB}
\title{Executar SQL em SGBD}
\usage{
runSQLonDB(nomeConexaoODBC, nomeArquivoComSQL = NULL, querySQL = NULL,
  nomeBanco = NULL, padCPFeCNPJ = TRUE, showSuccessMessage = TRUE)
}
\arguments{
\item{nomeConexaoODBC}{nome do Data Source configurado para o SGBD desejado.}

\item{nomeArquivoComSQL}{nome do arquivo .sql a ser executado. Só executa caso não
haja objeto querySQL definido.}

\item{querySQL}{objeto de texto/character com comando SQL a ser executado.}

\item{nomeBanco}{nome do banco de dados do SGBD onde dados serão consultados.}

\item{padCPFeCNPJ}{flag que define se deve haver padding de colunas de CPF e/ou CNPJ.}

\item{showSuccessMessage}{flag que define se a mensagem de sucesso deverá ser exibida.}
}
\value{
um data frame com resultado do SQL.
}
\description{
Esta função se conecta a um SGBD definido por um Data Source configurado e 
usa usuário/senha obtidos em arquivo database_logins.txt e executa SQL passado
via objeto ou arquivo.
}
\details{
Esta função inicialmente lê o arquivo database_logins.txt que deve estar na
mesma pasta do Working Directory (definido com setwd) e obtém usuário e senha 
para o Data Source indicado como argumento. 
Em seguida, havendo objeto com SQL em querySQL, o mesmo é executado, caso contrário
executa o arquivo .sql em nomeArquivoComSQL.
Para flag padCPFeCNPJ setada, o resultado do SQL é tratado para realizar padding de
qualquer coluna contendo o texto cpf/CPF/cnpj/CNPJ.
Para flag showSuccessMessage setada, havendo sucesso na consulta, a mensagem
de sucesso padrão é exibida. Caso sejam executados vários SQLs em sequência,
recomenda-se definir showSuccessMessage como FALSE, pois assim somente aparece
mensagem em caso de ERRO em algum SQL.
}
\examples{
# Obter dados do Data Source Esfinge configurado
# SQL encontra-se no arquivo de nome "Query_Esfinge_GABCRG_Punidos.sql"
dadosESFINGE <- runSQLonDB("Esfinge", "Query_Esfinge_GABCRG_Punidos.sql")
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{insertDataIntoDB}, \code{instrucoes.SQLServer}, \code{instrucoes.MySQL}
}

