% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARA.R
\name{updateDataInDB}
\alias{updateDataInDB}
\title{Atualizar dados em tabela de SGBD}
\usage{
updateDataInDB(nomeConexaoODBC, dadosATUALIZAR, nomeBanco = NULL, nomeTabela,
  chavePrimaria, showSuccessMessage = TRUE)
}
\arguments{
\item{nomeConexaoODBC}{nome do Data Source configurado para o SGBD desejado.}

\item{dadosATUALIZAR}{data frame com dados a serem inseridos. Deve conter colunas com
exatamente os mesmos nomes que as colunas da tabela onde dados serão inseridos 
(case sensitive).}

\item{nomeBanco}{nome do banco de dados do SGBD onde dados serão inseridos.}

\item{nomeTabela}{nome da tabela do SGBD onde dados serão inseridos.}

\item{chavePrimaria}{flag que define se deve exibir mensagens de inserção do SGBD.}

\item{showSuccessMessage}{flag que define se a mensagem de sucesso deverá ser exibida.}
}
\value{
mensagem de erro ou sucesso.
}
\description{
Esta função se conecta a um SGBD definido por um Data Source configurado e 
usa usuário/senha obtidos em arquivo database_logins.txt e atualiza dados passados
na tabela escolhida considerando a chave indicada.
}
\details{
Esta função inicialmente lê o arquivo database_logins.txt que deve estar na
mesma pasta do Working Directory (definido com setwd) e obtém usuário e senha 
para o Data Source indicado como argumento. 
Em seguida, atualiza dados do data frame dadosATUALIZAR na tabela nomeTabela.
O objeto dadosATUALIZAR deve conter colunas com exatamente os mesmos nomes que as colunas 
da tabela onde dados serão atualizados (case sensitive).
Deve-se indicar ainda qual o campo considerado chave primária através do argumento chavePrimaria.
Para flag showSuccessMessage setada, havendo sucesso na consulta, a mensagem
de sucesso padrão é exibida.
}
\examples{
# Atualizar dados de data frame df_RFB usando data source sed-die-bd1-c
# na tabela dm_rfb.rfb localizada no banco dw_mara_stage com a chave primária nr_cpf
updateDataInDB('sed-die-bd1-c', df_RFB, 'dw_mara_stage', 'dm_rfb.rfb', 'nr_cpf')
}
\author{
Ricardo S. Carvalho
}
\seealso{
\code{runSQLonDB}, \code{insertDataIntoDB}, \code{instrucoes.SQLServer}, \code{instrucoes.MySQL}
}

